[v1.0] Move domain-specific model builder into DENAS
