+++ e2eAIOK/DeNas/asr/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ['config','data','decoders','lib','scripts','trainer','utils','asr_model_builder','asr_trainer','init_asr_parser','supernet_asr','train','TransformerBase','TransformerLM']
+__all__ = ['config','data','decoders','lib','scripts','trainer','utils','asr_model_builder','asr_trainer','init_asr_parser','supernet_asr','train','TransformerBase','TransformerLM', 'model_builder_asr']
