+++ e2eAIOK/DeNas/cv/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ['third_party','utils','benchmark_network_latency','DeNet','supernet_transformer','cv_trainer']
+__all__ = ['third_party','utils','benchmark_network_latency','DeNet','supernet_transformer','cv_trainer', 'model_builder_cv']
