+++ e2eAIOK/DeNas/nlp/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ['supernet_bert','supernet_train','utils']
+__all__ = ['supernet_bert','supernet_train','utils', 'model_builder_nlp']
