+++ e2eAIOK/DeNas/train.py
@@ -9,9 +9,9 @@ import sentencepiece as sp
 import e2eAIOK.common.trainer.utils.extend_distributed as ext_dist
 from e2eAIOK.common.trainer.torch_trainer import TorchTrainer 
 import e2eAIOK.common.trainer.utils.utils as utils
-from e2eAIOK.common.trainer.model.model_builder_asr import ModelBuilderASR
-from e2eAIOK.common.trainer.model.model_builder_cv import ModelBuilderCV
-from e2eAIOK.common.trainer.model.model_builder_nlp import ModelBuilderNLP
+from asr.model_builder_asr import ModelBuilderASR
+from cv.model_builder_cv import ModelBuilderCV
+from nlp.model_builder_nlp import ModelBuilderNLP
 from e2eAIOK.common.trainer.data.data_builder_librispeech import DataBuilderLibriSpeech
 from e2eAIOK.common.trainer.data.data_builder_cifar import DataBuilderCIFAR
 from e2eAIOK.common.trainer.data.data_builder_nlp import DataBuilderNLP
