--- e2eAIOK/common/trainer/model/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ['model_builder_asr','model_builder_cv','model_builder_nlp']
